requirements.txt
pillow~=9.4.0
torch~=1.13.1+cu117
torchvision~=0.14.1+cu117
numpy~=1.24.3
timm~=0.6.12
scipy~=1.10.1
setuptools~=68.0.0
tqdm~=4.66.1
pretrained-backbones-unet